chore: split in-memory-storage into its own export to avoid importing compass specific stuff from storage
